+++ changelog.md
@@ -60,6 +60,8 @@ All notable changes to this project will be documented in this file.
 
 The following types are now supported on JVM, JS and Native platforms:
 
+- `kotools.types.number.Number`
+  ([#99](https://github.com/kotools/types/issues/99))
 - `KotoolsInt` ([#83](https://github.com/kotools/types/issues/83))
 - `NonZeroInt` ([#84](https://github.com/kotools/types/issues/84))
 - `NotBlankString` ([#89](https://github.com/kotools/types/issues/89)).